prostapka/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.6.3
click==8.1.7
Django==4.2.6
django-bootstrap5==23.3
django-crispy-forms==2.0
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
Pillow==10.0.1
protobuf==4.21.12
reportlab==4.0.6
soupsieve==2.5
typing_extensions==4.8.0
Werkzeug==3.0.0
